git-svn-id: https://projects.coin-or.org/svn/ADOL-C/stable/2.5@473 94ac48a7-3327-4b6a-8511-9a4036a20e83

+++ ADOL-C/examples/additional_examples/lighthouse/cubic-2.cpp
@@ -16,7 +16,7 @@
 
 /****************************************************************************/
 /*                                                                 INCLUDES */
-#include <adolc.h>
+#include <adolc/adolc.h>
 
 #include <math.h>
 #define PI 3.1415926536
@@ -88,7 +88,7 @@ int main() {
     trace_off();
 
     /*--------------------------------------------------------------------------*/
-    int tape_stats[STAT_SIZE];
+    size_t tape_stats[STAT_SIZE];
     tapestats(tag,tape_stats);
 
     fprintf(stdout,"\n    independents            %d\n",tape_stats[NUM_INDEPENDENTS]);
