+++ ADOL-C/examples/additional_examples/ode/vgenodemain.cpp
@@ -28,15 +28,15 @@
 
 /****************************************************************************/
 /*                                                                 INCLUDES */
-#include <adolc.h>
-#include <../examples/additional_examples/clock/myclock.h>
+#include <adolc/adolc.h>
+#include "../clock/myclock.h"
 
 #include <math.h>
 
 
 /****************************************************************************/
 /*                                                                   MACROS */
-#define TIMEFORMAT " %12.6le units,   %12.6le scunits,   %12.6le seconds\n"
+#define TIMEFORMAT " %12.6E units,   %12.6E scunits,   %12.6E seconds\n"
 
 
 /****************************************************************************/
@@ -154,7 +154,7 @@ int main() {
     tapingVectorFunction(tag,indeps,deps);
     t01 = myclock();
 
-    int tape_stats[STAT_SIZE];
+    size_t tape_stats[STAT_SIZE];
     tapestats(tag,tape_stats);
 
     fprintf(stdout,"\n    independents            %d\n",tape_stats[NUM_INDEPENDENTS]);
@@ -253,7 +253,7 @@ int main() {
                 avg = 0.1;
             err += fabs(w[j][i]*tau-indeps2[j][i+1]*(i+1))/avg;
         }
-    fprintf(stdout,"%14.6le = total error \n",err);
+    fprintf(stdout,"%14.6E = total error \n",err);
 
     /*------------------------------------------------------------------------*/
     /* If desired print out Jacobians of Taylor coeffcients with
@@ -265,7 +265,7 @@ int main() {
             fprintf(stdout,"\n\t<-- B(%d)\n",i);
             for (j=0; j<indepDim; j++) {
                 for (k=0; k<indepDim; k++)
-                    fprintf(stdout,"%14.6le ",B[j][k][i]);
+                    fprintf(stdout,"%14.6E ",B[j][k][i]);
                 fprintf(stdout,"\n");
             }
         }
@@ -289,7 +289,7 @@ int main() {
                            fabs(1-(w[j][i+1]-indeps2[j][i+1])/h/B[j][k][i])
                            : fabs((w[j][i+1]-indeps2[j][i+1])/h) ;
             }
-            fprintf(stdout," Relative truncation errors in B(%d) ---> %14.6le\n",
+            fprintf(stdout," Relative truncation errors in B(%d) ---> %14.6E\n",
                     i,err);
         }
